requirements.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
argh==0.26.2
asn1crypto @ file:///D:/bld/asn1crypto_1588618201491/work
astroid @ file:///D:/bld/astroid_1588698401017/work
astropy==4.0.1.post1
atomicwrites @ file:///home/conda/feedstock_root/build_artifacts/atomicwrites_1588182545583/work
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bitarray==1.2.1
bkcharts==0.2
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1588608214987/work
bokeh==2.0.1
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1588164361239/work
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.8.3
conda-build==3.19.2
conda-package-handling==1.6.0
conda-verify==3.1.1
contextlib2==0.6.0.post1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1588975274559/work
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distributed @ file:///D:/bld/distributed_1588976268783/work
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.10
flake8 @ file:///home/conda/feedstock_root/build_artifacts/flake8_1589343427411/work
Flask==1.1.2
fsspec==0.7.3
future==0.18.2
gevent @ file:///D:/bld/gevent_1588347569753/work
gitdb2==2.0.6
githubcommit==0.1.0
GitPython==3.0.5
glob2==0.7
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
greenlet==0.4.15
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.17.0
hypothesis @ file:///home/conda/feedstock_root/build_artifacts/hypothesis_1589415572283/work
idna==2.9
imagecodecs-lite==2019.12.3
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.2.1
ipython @ file:///D:/bld/ipython_1588363073263/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.11.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1589548914243/work
json5==0.9.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.1.2
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1588946599005/work
jupyterthemes==0.20.0
keyring @ file:///D:/bld/keyring_1588434024659/work
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lesscpy==0.13.0
libarchive-c==2.9
lightgbm==2.3.1
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
mock @ file:///D:/bld/mock_1588619039489/work
more-itertools @ file:///home/conda/feedstock_root/build_artifacts/more-itertools_1589614090069/work
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.4.4
nose==1.3.7
notebook @ file:///D:/bld/notebook_1588887315004/work
numba==0.48.0
numexpr==2.7.1
numpy @ file:///D:/bld/numpy_1588605187257/work
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
packaging==20.1
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy @ file:///D:/bld/pluggy_1588616654863/work
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1589305246696/work
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylint @ file:///D:/bld/pylint_1588950636619/work
Pympler==0.8
PyNaCl==1.3.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.2
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.2
pytest-doctestplus @ file:///home/conda/feedstock_root/build_artifacts/pytest-doctestplus_1588699114696/work
pytest-openfiles==0.4.0
pytest-remotedata==0.3.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.2
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
qgrid==1.1.1
QtAwesome @ file:///home/conda/feedstock_root/build_artifacts/qtawesome_1588715536397/work
qtconsole @ file:///home/conda/feedstock_root/build_artifacts/qtconsole_1589392299544/work
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rise==5.6.1
rope @ file:///home/conda/feedstock_root/build_artifacts/rope_1588950285934/work
rsa==4.0
Rtree==0.9.4
ruamel-yaml==0.15.80
scikit-image==0.17.2
scikit-learn @ file:///D:/bld/scikit-learn_1589375624451/work
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.22
six==1.14.0
smmap2==2.0.5
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.4
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-websupport_1588158706257/work
spyder==4.0.0
spyder-kernels @ file:///D:/bld/spyder-kernels_1588785911039/work
SQLAlchemy @ file:///D:/bld/sqlalchemy_1589421283037/work
statsmodels==0.11.1
swat==1.6.1
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1589323165428/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1589469402899/work
toolz==0.10.0
tornado==6.0.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1588513834538/work
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1588470653596/work
ujson==2.0.3
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.19.2
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.29.0
zict==2.0.0
zipp==3.1.0
